+++ src/components/BookList.js
@@ -1,19 +1,19 @@
-import React from "react";
+import React, { useContext } from "react";
+
+import { BooksContext } from "../context/contextBooks";
 
 import "../styles/index.css";
 
 import BookItem from "./BookItem";
 
-const BookList = ({ numberOfBooks, deleteBookById, editBookById }) => {
+const BookList = () => {
+  const { numberOfBooks } = useContext(BooksContext);
+
   return (
     <div className="book-list">
       {numberOfBooks.map((book) => (
         <div key={book.id}>
-          <BookItem
-            editBookById={editBookById}
-            deleteBookById={deleteBookById}
-            book={book}
-          />
+          <BookItem book={book} />
         </div>
       ))}
     </div>
