--- src/components/EditBook.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const EditBook = () => {
-  return <div>EditBook</div>;
-};
-
-export default EditBook;
